+++ jvm/assemble.sh
@@ -41,6 +41,10 @@ cp "${TARGET_DIR}/armv7-unknown-linux-musleabihf/release/libnostr_sdk_ffi.so" "$
 cp "${TARGET_DIR}/aarch64-unknown-linux-musl/release/libnostr_sdk_ffi.so" "${RESOURCE_DIR}/linux-aarch64-musl/"
 cp "${TARGET_DIR}/riscv64gc-unknown-linux-musl/release/libnostr_sdk_ffi.so" "${RESOURCE_DIR}/linux-riscv64-musl/"
 
+# Copy FreeBSD binaries
+mkdir -p "${RESOURCE_DIR}/freebsd-x86-64/"
+cp "${TARGET_DIR}/x86_64-unknown-freebsd/release/libnostr_sdk_ffi.so" "${RESOURCE_DIR}/freebsd-x86-64/"
+
 # Copy windows binaries
 mkdir -p "${RESOURCE_DIR}/win32-x86/"
 mkdir -p "${RESOURCE_DIR}/win32-x86-64/"