+++ scripts/all.sh
@@ -7,6 +7,7 @@ set -exuo pipefail
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
 "${SCRIPT_DIR}/android.sh"
+"${SCRIPT_DIR}/freebsd.sh"
 "${SCRIPT_DIR}/linux.sh"
 "${SCRIPT_DIR}/macos.sh"
 "${SCRIPT_DIR}/windows.sh"